src/main/java/be/giantoaf/recipebook/service/CategoryService.java
package be.giantoaf.recipebook.service;

import be.giantoaf.recipebook.model.Category;
import be.giantoaf.recipebook.repository.CategoryRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class CategoryService {

    private final CategoryRepository categoryRepository;

    public CategoryService(CategoryRepository categoryRepository) {
        this.categoryRepository = categoryRepository;
    }

    public List<Category> getAllCategories() {
        return categoryRepository.findAll();
    }

}
